% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{additionalpar}
\alias{additionalpar}
\title{AdditionalPar}
\format{
An object of class \code{data.frame} with 78 rows and 12 columns.
}
\usage{
additionalpar
}
\description{
National addtitional parameter for classification (regions, special building types etc.)
}
\keyword{datasets}
