% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atticcond}
\alias{atticcond}
\title{AtticCond}
\format{
An object of class \code{data.frame} with 6 rows and 5 columns.
}
\usage{
atticcond
}
\description{
Space heating situation of the attic storey
}
\keyword{datasets}
