% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attneighb}
\alias{attneighb}
\title{AttNeighb}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
attneighb
}
\description{
Codes for the number of attached neighbour buildings
}
\keyword{datasets}
