% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundarycond}
\alias{boundarycond}
\title{BoundaryCond}
\format{
An object of class \code{data.frame} with 62 rows and 38 columns.
}
\usage{
boundarycond
}
\description{
Boundary conditions for the energy balance calculation (optional)
}
\keyword{datasets}
