% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boundarycondtype}
\alias{boundarycondtype}
\title{BoundaryCondType}
\format{
An object of class \code{data.frame} with 2 rows and 2 columns.
}
\usage{
boundarycondtype
}
\description{
Type of boundary condition input
}
\keyword{datasets}
