% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{building.constr}
\alias{building.constr}
\title{Building.Constr}
\format{
An object of class \code{data.frame} with 2344 rows and 35 columns.
}
\usage{
building.constr
}
\description{
National definition of construction elements + U-values
}
\keyword{datasets}
