% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{building.measure}
\alias{building.measure}
\title{Building.Measure}
\format{
An object of class \code{data.frame} with 1300 rows and 78 columns.
}
\usage{
building.measure
}
\description{
National definition of insulation measures + thermal resistance
}
\keyword{datasets}
