% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buildingsizeclass}
\alias{buildingsizeclass}
\title{BuildingSizeClass}
\format{
An object of class \code{data.frame} with 5 rows and 5 columns.

An object of class \code{data.frame} with 5 rows and 5 columns.
}
\usage{
buildingsizeclass

buildingsizeclass
}
\description{
Building size classes

Building size classes
}
\keyword{datasets}
