% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calcadapt}
\alias{calcadapt}
\title{CalcAdapt}
\format{
An object of class \code{data.frame} with 49 rows and 20 columns.
}
\usage{
calcadapt
}
\description{
Definition of calibration factors for adapting the TABULA calculation results to the typical level of metered consumption or to the typical level of national calculation methods (used for respective estimations)
}
\keyword{datasets}
