% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellarcond}
\alias{cellarcond}
\title{CellarCond}
\format{
An object of class \code{data.frame} with 6 rows and 5 columns.
}
\usage{
cellarcond
}
\description{
Space heating situation of the cellar storey
}
\keyword{datasets}
