% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate}
\alias{climate}
\title{Climate}
\format{
An object of class \code{data.frame} with 74 rows and 27 columns.
}
\usage{
climate
}
\description{
National and regional climate conditions
}
\keyword{datasets}
