% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complexfootprint}
\alias{complexfootprint}
\title{ComplexFootprint}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
complexfootprint
}
\description{
Degree of geometrical complexity of the footprint shape
}
\keyword{datasets}
