% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.atticcond}
\alias{const.atticcond}
\title{Const.AtticCond}
\format{
An object of class \code{data.frame} with 6 rows and 5 columns.
}
\usage{
const.atticcond
}
\description{
Space heating situation of the attic storey
}
\keyword{datasets}
