% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.attneighb}
\alias{const.attneighb}
\title{Const.AttNeighb}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
const.attneighb
}
\description{
Codes for the number of attached neighbour buildings
}
\keyword{datasets}
