% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.buildingsizeclass}
\alias{const.buildingsizeclass}
\title{Const.BuildingSizeClass}
\format{
An object of class \code{data.frame} with 5 rows and 5 columns.
}
\usage{
const.buildingsizeclass
}
\description{
Building size classes
}
\keyword{datasets}
