% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.cellarcond}
\alias{const.cellarcond}
\title{Const.CellarCond}
\format{
An object of class \code{data.frame} with 6 rows and 5 columns.
}
\usage{
const.cellarcond
}
\description{
Space heating situation of the cellar storey
}
\keyword{datasets}
