% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.complexfootprint}
\alias{const.complexfootprint}
\title{Const.ComplexFootprint}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
const.complexfootprint
}
\description{
Degree of geometrical complexity of the footprint shape
}
\keyword{datasets}
