% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.complexroof}
\alias{const.complexroof}
\title{Const.ComplexRoof}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
const.complexroof
}
\description{
Degree of geometrical complexity of the roof shape
}
\keyword{datasets}
