% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.country}
\alias{const.country}
\title{Const.Country}
\format{
An object of class \code{data.frame} with 52 rows and 5 columns.
}
\usage{
const.country
}
\description{
Country codes
}
\keyword{datasets}
