% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.dataformat}
\alias{const.dataformat}
\title{Const.DataFormat}
\format{
An object of class \code{data.frame} with 8 rows and 3 columns.
}
\usage{
const.dataformat
}
\description{
Data format codes
}
\keyword{datasets}
