% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.elementtype}
\alias{const.elementtype}
\title{Const.ElementType}
\format{
An object of class \code{data.frame} with 7 rows and 25 columns.
}
\usage{
const.elementtype
}
\description{
Thermal envelope element types
}
\keyword{datasets}
