% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.energycarrier}
\alias{const.energycarrier}
\title{Const.EnergyCarrier}
\format{
An object of class \code{data.frame} with 59 rows and 50 columns.
}
\usage{
const.energycarrier
}
\description{
Codes for the used energy carriers
}
\keyword{datasets}
