% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.language}
\alias{const.language}
\title{Const.Language}
\format{
An object of class \code{data.frame} with 23 rows and 5 columns.
}
\usage{
const.language
}
\description{
Language codes
}
\keyword{datasets}
