% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.measuretype}
\alias{const.measuretype}
\title{Const.MeasureType}
\format{
An object of class \code{data.frame} with 3 rows and 25 columns.
}
\usage{
const.measuretype
}
\description{
Type of refurbishment measure / replacement of existing insulation or elements
}
\keyword{datasets}
