% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.orientation}
\alias{const.orientation}
\title{Const.Orientation}
\format{
An object of class \code{data.frame} with 9 rows and 3 columns.
}
\usage{
const.orientation
}
\description{
Geographical orientation
}
\keyword{datasets}
