% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.rooftype}
\alias{const.rooftype}
\title{Const.RoofType}
\format{
An object of class \code{data.frame} with 3 rows and 5 columns.
}
\usage{
const.rooftype
}
\description{
Roof type codes
}
\keyword{datasets}
