% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.statusdataset}
\alias{const.statusdataset}
\title{Const.StatusDataset}
\format{
An object of class \code{data.frame} with 8 rows and 3 columns.
}
\usage{
const.statusdataset
}
\description{
Dataset status codes
}
\keyword{datasets}
