% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.systype.ha}
\alias{const.systype.ha}
\title{Const.SysType.HA}
\format{
An object of class \code{data.frame} with 2 rows and 5 columns.
}
\usage{
const.systype.ha
}
\description{
Heating system: auxiliary energy types
}
\keyword{datasets}
