% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.systype.hd}
\alias{const.systype.hd}
\title{Const.SysType.HD}
\format{
An object of class \code{data.frame} with 9 rows and 5 columns.
}
\usage{
const.systype.hd
}
\description{
Heating system: distribution and heat emission types
}
\keyword{datasets}
