% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.systype.hg}
\alias{const.systype.hg}
\title{Const.SysType.HG}
\format{
An object of class \code{data.frame} with 32 rows and 5 columns.
}
\usage{
const.systype.hg
}
\description{
Heating system: generator types
}
\keyword{datasets}
