% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.systype.hs}
\alias{const.systype.hs}
\title{Const.SysType.HS}
\format{
An object of class \code{data.frame} with 4 rows and 5 columns.
}
\usage{
const.systype.hs
}
\description{
Heating system: storage types
}
\keyword{datasets}
