% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.systype.size}
\alias{const.systype.size}
\title{Const.SysType.Size}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
const.systype.size
}
\description{
Building types used to distinguish between different system component sizes
}
\keyword{datasets}
