% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.systype.wg}
\alias{const.systype.wg}
\title{Const.SysType.WG}
\format{
An object of class \code{data.frame} with 25 rows and 5 columns.
}
\usage{
const.systype.wg
}
\description{
Domestic hot water system: generator types
}
\keyword{datasets}
