% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.thermalbridging}
\alias{const.thermalbridging}
\title{Const.ThermalBridging}
\format{
An object of class \code{data.frame} with 5 rows and 6 columns.
}
\usage{
const.thermalbridging
}
\description{
Classification of thermal bridging
}
\keyword{datasets}
