% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{const.utilisation}
\alias{const.utilisation}
\title{Const.Utilisation}
\format{
An object of class \code{data.frame} with 33 rows and 26 columns.
}
\usage{
const.utilisation
}
\description{
Common definitions of utilisation types
}
\keyword{datasets}
