% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constrborder}
\alias{constrborder}
\title{ConstrBorder}
\format{
An object of class \code{data.frame} with 4 rows and 7 columns.
}
\usage{
constrborder
}
\description{
Type of construction border situation / location of construction element
}
\keyword{datasets}
