% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constryearclass}
\alias{constryearclass}
\title{ConstrYearClass}
\format{
An object of class \code{data.frame} with 155 rows and 13 columns.
}
\usage{
constryearclass
}
\description{
National construction year classes
}
\keyword{datasets}
