% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datatype.building}
\alias{datatype.building}
\title{DataType.Building}
\format{
An object of class \code{data.frame} with 4 rows and 3 columns.
}
\usage{
datatype.building
}
\description{
Building data type
}
\keyword{datasets}
