% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elementtype}
\alias{elementtype}
\title{ElementType}
\format{
An object of class \code{data.frame} with 7 rows and 25 columns.
}
\usage{
elementtype
}
\description{
Thermal envelope element types
}
\keyword{datasets}
