% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energycarrier}
\alias{energycarrier}
\title{EnergyCarrier}
\format{
An object of class \code{data.frame} with 55 rows and 50 columns.
}
\usage{
energycarrier
}
\description{
Codes for the used energy carriers
}
\keyword{datasets}
