% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infiltration}
\alias{infiltration}
\title{Infiltration}
\format{
An object of class \code{data.frame} with 4 rows and 6 columns.
}
\usage{
infiltration
}
\description{
Classification of infiltration, dependent on air tightness
}
\keyword{datasets}
