% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info.sheet}
\alias{info.sheet}
\title{Info.Sheet}
\format{
An object of class \code{data.frame} with 69 rows and 7 columns.
}
\usage{
info.sheet
}
\description{
Internal sheet control
}
\keyword{datasets}
