% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insulation.default}
\alias{insulation.default}
\title{Insulation.Default}
\format{
An object of class \code{data.frame} with 41 rows and 29 columns.
}
\usage{
insulation.default
}
\description{
National default values of insulation thickness and thermal conductivity of insulation measures, differentiated by implementation period and construction type; to be used if the actual values are not known for a given construction (usage of these default values increases the uncertainty)
}
\keyword{datasets}
