% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure.f.default}
\alias{measure.f.default}
\title{Measure.f.Default}
\format{
An object of class \code{data.frame} with 76 rows and 18 columns.
}
\usage{
measure.f.default
}
\description{
National default values of insulation fraction, differentiated by construction type, implementation period; two cases: (a) typical insulation fraction of refurbished construtions (indicator "Refurbished", used if actual insulation fraction is not known for a construction); (b) ratio of insulated to total construction area for the building stock of the specific construction period (indicator "NA", values used to model an average state for constructions of unknown state) - usage of these default values increases the uncertainty
}
\keyword{datasets}
