% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meter.energydensity}
\alias{meter.energydensity}
\title{TODO}
\format{
An object of class \code{data.frame} with 27 rows and 76 columns.
}
\usage{
meter.energydensity
}
\description{
yyyy
}
\keyword{datasets}
