% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par.envareaestim}
\alias{par.envareaestim}
\title{Par.EnvAreaEstim}
\format{
An object of class \code{data.frame} with 2 rows and 42 columns.
}
\usage{
par.envareaestim
}
\description{
Parameters for the envelope area estimation procedure
}
\keyword{datasets}
