% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statusdataset}
\alias{statusdataset}
\title{StatusDataset}
\format{
An object of class \code{data.frame} with 8 rows and 3 columns.
}
\usage{
statusdataset
}
\description{
Dataset status codes
}
\keyword{datasets}
