% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.coverage}
\alias{system.coverage}
\title{System.Coverage}
\format{
An object of class \code{data.frame} with 49 rows and 34 columns.
}
\usage{
system.coverage
}
\description{
Default values of coverages depending on supply/load ratio
}
\keyword{datasets}
