% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.ec}
\alias{system.ec}
\title{System.EC}
\format{
An object of class \code{data.frame} with 308 rows and 25 columns.
}
\usage{
system.ec
}
\description{
Energy carrier specification and assessment
}
\keyword{datasets}
