% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.elprod}
\alias{system.elprod}
\title{System.ElProd}
\format{
An object of class \code{data.frame} with 67 rows and 24 columns.
}
\usage{
system.elprod
}
\description{
Assessment of on-site produced electricity
}
\keyword{datasets}
