% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.h}
\alias{system.h}
\title{System.H}
\format{
An object of class \code{data.frame} with 666 rows and 31 columns.
}
\usage{
system.h
}
\description{
National configurations of heating system
}
\keyword{datasets}
