% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.ha}
\alias{system.ha}
\title{System.HA}
\format{
An object of class \code{data.frame} with 117 rows and 15 columns.
}
\usage{
system.ha
}
\description{
Heating system / auxiliary energy default values
}
\keyword{datasets}
