% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.hd}
\alias{system.hd}
\title{System.HD}
\format{
An object of class \code{data.frame} with 214 rows and 15 columns.
}
\usage{
system.hd
}
\description{
Heating system / heat distribution default values
}
\keyword{datasets}
