% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.hg}
\alias{system.hg}
\title{System.HG}
\format{
An object of class \code{data.frame} with 620 rows and 16 columns.
}
\usage{
system.hg
}
\description{
Heating system / heat generation default values
}
\keyword{datasets}
