% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.hs}
\alias{system.hs}
\title{System.HS}
\format{
An object of class \code{data.frame} with 102 rows and 15 columns.

An object of class \code{data.frame} with 102 rows and 15 columns.
}
\usage{
system.hs

system.hs
}
\description{
Heating system / heat storage default values

Heating system / heat storage default values
}
\keyword{datasets}
