% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.pv}
\alias{system.pv}
\title{System.PV}
\format{
An object of class \code{data.frame} with 161 rows and 32 columns.
}
\usage{
system.pv
}
\description{
Default values of PV electricity production depending on orientation
}
\keyword{datasets}
