% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.pvpanel}
\alias{system.pvpanel}
\title{System.PVPanel}
\format{
An object of class \code{data.frame} with 34 rows and 18 columns.
}
\usage{
system.pvpanel
}
\description{
Default values of PV electricity production depending on orientation
}
\keyword{datasets}
