% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.setecassess}
\alias{system.setecassess}
\title{System.SetECAssess}
\format{
An object of class \code{data.frame} with 26 rows and 12 columns.
}
\usage{
system.setecassess
}
\description{
Definition of energy carrier assessment sets
}
\keyword{datasets}
