% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.w}
\alias{system.w}
\title{System.W}
\format{
An object of class \code{data.frame} with 517 rows and 31 columns.
}
\usage{
system.w
}
\description{
National configurations of domestic hot water systems
}
\keyword{datasets}
