% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.wa}
\alias{system.wa}
\title{System.WA}
\format{
An object of class \code{data.frame} with 127 rows and 15 columns.
}
\usage{
system.wa
}
\description{
Domestic hot water system / auxiliary energy default values
}
\keyword{datasets}
