% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.wd}
\alias{system.wd}
\title{System.WD}
\format{
An object of class \code{data.frame} with 217 rows and 16 columns.
}
\usage{
system.wd
}
\description{
Domestic hot water system / heat distribution default values
}
\keyword{datasets}
