% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.wg}
\alias{system.wg}
\title{System.WG}
\format{
An object of class \code{data.frame} with 516 rows and 16 columns.
}
\usage{
system.wg
}
\description{
Domestic hot water system / heat generation default values
}
\keyword{datasets}
