% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{system.ws}
\alias{system.ws}
\title{System.WS}
\format{
An object of class \code{data.frame} with 189 rows and 16 columns.
}
\usage{
system.ws
}
\description{
Domestic hot water system / heat storage default values
}
\keyword{datasets}
