% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{systype.vent}
\alias{systype.vent}
\title{SysType.Vent}
\format{
An object of class \code{data.frame} with 5 rows and 5 columns.
}
\usage{
systype.vent
}
\description{
Heating system: ventilation types
}
\keyword{datasets}
