% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{systype.wa}
\alias{systype.wa}
\title{SysType.WA}
\format{
An object of class \code{data.frame} with 5 rows and 5 columns.
}
\usage{
systype.wa
}
\description{
Domestic hot water system: auxiliary energy types
}
\keyword{datasets}
