% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{systype.wd}
\alias{systype.wd}
\title{SysType.WD}
\format{
An object of class \code{data.frame} with 8 rows and 5 columns.
}
\usage{
systype.wd
}
\description{
Domestic hot water system: distribution types
}
\keyword{datasets}
