% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{systype.ws}
\alias{systype.ws}
\title{SysType.WS}
\format{
An object of class \code{data.frame} with 6 rows and 5 columns.
}
\usage{
systype.ws
}
\description{
Domestic hot water system: storage types
}
\keyword{datasets}
