% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thermalbridging}
\alias{thermalbridging}
\title{ThermalBridging}
\format{
An object of class \code{data.frame} with 5 rows and 6 columns.
}
\usage{
thermalbridging
}
\description{
Classification of thermal bridging
}
\keyword{datasets}
