% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{type.calcadapt}
\alias{type.calcadapt}
\title{Type.CalcAdapt}
\format{
An object of class \code{data.frame} with 2 rows and 5 columns.
}
\usage{
type.calcadapt
}
\description{
Calculation adaptation types
}
\keyword{datasets}
