% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typeintake}
\alias{typeintake}
\title{TypeIntake}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\usage{
typeintake
}
\description{
Types of data intake
}
\keyword{datasets}
