% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typevariant}
\alias{typevariant}
\title{TypeVariant}
\format{
An object of class \code{data.frame} with 2 rows and 3 columns.
}
\usage{
typevariant
}
\description{
Type of building variant (refurbishment variant or new build variant)
}
\keyword{datasets}
