% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typologyregion}
\alias{typologyregion}
\title{TypologyRegion}
\format{
An object of class \code{data.frame} with 57 rows and 10 columns.
}
\usage{
typologyregion
}
\description{
National typology regions
}
\keyword{datasets}
