% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u.class.constr}
\alias{u.class.constr}
\title{U.Class.Constr}
\format{
An object of class \code{data.frame} with 123 rows and 34 columns.
}
\usage{
u.class.constr
}
\description{
National default U-values differentiated by construction year class and construction type
}
\keyword{datasets}
