% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u.class.window}
\alias{u.class.window}
\title{U.Class.Window}
\format{
An object of class \code{data.frame} with 42 rows and 22 columns.
}
\usage{
u.class.window
}
\description{
U-values of windows (old scheme used for TABULA WebTool
(only three time bands))
}
\keyword{datasets}
